refactor: FieldType 패키지 이동
