+++ src/main/java/persistence/sql/FieldType.java
@@ -1,4 +1,4 @@
-package persistence.sql.ddl;
+package persistence.sql;
 
 import java.lang.reflect.Field;
