+++ src/test/java/persistence/sql/FieldTypeTest.java
@@ -1,4 +1,4 @@
-package persistence.sql.ddl;
+package persistence.sql;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
